+++ src/c2/win/WindowsToolOutputParseException.java
@@ -1,0 +1,13 @@
+package c2.win;
+
+public class WindowsToolOutputParseException extends Exception{
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+
+	public WindowsToolOutputParseException(String message) {
+		super(message);
+	}
+}